# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A high-performance REST API service for scraping Google search results using SerpAPI. Designed for lead generation workflows.",
        "title": "Lead Gen Worker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/webhooks/job-created": {
            "post": {
                "description": "Receives Supabase database webhook when a new job is created",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Handle job created webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token with webhook secret",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Job payload from Supabase",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.Job"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "webstar_noturno-leadgen-worker_internal_dto.ErrorResponse": {
            "description": "Error response returned when request fails",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message describing what went wrong",
                    "type": "string",
                    "example": "Key: 'SearchRequest.Q' Error:Field validation for 'Q' failed on the 'required' tag"
                }
            }
        },
        "webstar_noturno-leadgen-worker_internal_dto.Job": {
            "type": "object",
            "properties": {
                "business_profile": {
                    "description": "ID of the business profile to use for personalization",
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "excluded_domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "icp_id": {
                    "type": "string"
                },
                "icp_name": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "lead_quantity": {
                    "type": "integer"
                },
                "leads_generated": {
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "required_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, processing, completed, failed",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}